Struct ncollide_geometry::bounding_volume::AABB

pub struct AABB<P> { /* fields omitted */ }

An Axis Aligned Bounding Box.

Methods

impl<P: Point> AABB<P>

fn new(mins: P, maxs: P) -> AABB<P>

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

impl<P: Point> AABB<P>

fn new_invalid() -> AABB<P>

Creates an invalid AABB with: * mins = Bounded::max_value() * maxs = Bounded::max_value(). This is useful to build aabb using merges.

impl<P> AABB<P>

fn mins(&self) -> &P

Reference to the AABB point with the smallest components along each axis.

fn maxs(&self) -> &P

Reference to the AABB point with the biggest components along each axis.

impl<P> AABB<P> where P: Point

fn center(&self) -> P

The center of this AABB.

fn half_extents(&self) -> P::Vect

The half extents of this AABB.