Struct ncollide_geometry::partitioning::DBVT

pub struct DBVT<P, B, BV> { /* fields omitted */ }

A Dynamic Bounding Volume Tree.

Methods

impl<P, B, BV> DBVT<P, B, BV>

fn new() -> DBVT<P, B, BV>

Creates a new Dynamic Bounding Volume Tree.

impl<P, B, BV> DBVT<P, B, BV> where P: Point,
BV: 'static + BoundingVolume<P> + Clone,
B: 'static + Clone

fn remove(&mut self, leaf: &mut Rc<RefCell<DBVTLeaf<P, B, BV>>>)

Removes a leaf from the tree. Fails if the tree is empty.

fn insert_new(&mut self, b: B, bv: BV) -> Rc<RefCell<DBVTLeaf<P, B, BV>>>

Creates, inserts, and returns a new leaf with the given content.

fn insert(&mut self, leaf: Rc<RefCell<DBVTLeaf<P, B, BV>>>)

Inserts a leaf to the tree.

fn visit<Vis: BVTVisitor<B, BV>>(&self, visitor: &mut Vis)

Traverses this tree using an object implementing the BVTVisitortrait.

This will traverse the whole tree and call the visitor .visit_internal(...) (resp. .visit_leaf(...)) method on each internal (resp. leaf) node.