Function ncollide_geometry::query::algorithms::minkowski_sampling::closest_points

pub fn closest_points<P, M, S, G1: ?Sized, G2: ?Sized>(m1: &M,
                                                       g1: &G1,
                                                       m2: &M,
                                                       g2: &G2,
                                                       simplex: &mut S)
                                                       -> Option<(P, P, P::Vect)> where P: Point,
        P::Vect: Translate<P>,
        M: Translation<P::Vect>,
        S: Simplex<AnnotatedPoint<P>>,
        G1: SupportMap<P, M>,
        G2: SupportMap<P, M>

Computes the closest points between two implicit inter-penetrating shapes. Returns None if the shapes are not in penetration. This can be used as a fallback algorithm for the GJK algorithm.