Struct ncollide_geometry::query::point_internal::PointProjection

pub struct PointProjection<P: Point> {
    pub is_inside: bool,
    pub point: P,
}

Description of the projection of a point on a shape.

Fields
is_inside: bool

Whether or not the point to project was inside of the shape.

point: P

The projection result.

Methods

impl<P: Point> PointProjection<P>

fn new(is_inside: bool, point: P) -> PointProjection<P>

Initializes a new PointProjection.