Trait ncollide_geometry::query::PointQuery

pub trait PointQuery<P: Point, M> {
    fn project_point(&self, m: &M, pt: &P, solid: bool) -> PointProjection<P>;

    fn distance_to_point(&self, m: &M, pt: &P, solid: bool) -> P::Real { ... }
    fn contains_point(&self, m: &M, pt: &P) -> bool { ... }
}

Trait of objects that can be tested for point inclusion and projection.

Required Methods

fn project_point(&self, m: &M, pt: &P, solid: bool) -> PointProjection<P>

Projects a point on self transformed by m.

Provided Methods

fn distance_to_point(&self, m: &M, pt: &P, solid: bool) -> P::Real

Computes the minimal distance between a point and self transformed by m.

fn contains_point(&self, m: &M, pt: &P) -> bool

Tests if the given point is inside of self transformed by m.