Struct ncollide_geometry::shape::Capsule

pub struct Capsule<N> { /* fields omitted */ }

SupportMap description of a capsule shape with its principal axis aligned with the y axis.

Methods

impl<N> Capsule<N> where N: Scalar

fn new(half_height: N, radius: N) -> Capsule<N>

Creates a new capsule.

Arguments:

half_height - the half length of the capsule along the y axis.

radius - radius of the rounded part of the capsule.

fn half_height(&self) -> N

The capsule half length along the y axis.

fn radius(&self) -> N

The radius of the capsule's rounded part.