Struct ncollide_geometry::shape::Plane

pub struct Plane<V> { /* fields omitted */ }

SupportMap description of a plane.

Methods

impl<V: Vector> Plane<V>

fn new(normal: V) -> Plane<V>

Builds a new plane from its center and its normal.

impl<V> Plane<V>

unsafe fn new_normalized(normal: V) -> Plane<V>

Builds a new plane from its center and its normal.

fn normal(&self) -> &V

The plane normal.